% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_sanger_data.R
\name{merge_sanger_data}
\alias{merge_sanger_data}
\title{Merge multiple sanger datasets together}
\usage{
merge_sanger_data(...)
}
\arguments{
\item{...}{Multiple base density datasets from simulate_sanger_data()}
}
\value{
A merged base density dataset
}
\description{
Merge multiple sanger datasets together
}
\examples{
merge_sanger_data(simulate_sanger_data("GAG"),simulate_sanger_data("GTG"))

}
