% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_documentation.R
\docType{package}
\name{sangr}
\alias{sangr}
\title{sangr: A package for simulating sanger sequencing chromatograms}
\description{
The sangr package provides three main functions which you might
want to use.
}
\section{sangr functions}{

simulate_sanger_data

draw_chromatogram

merge_sanger_data
}

