% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_sanger_data.R
\name{simulate_sanger_data}
\alias{simulate_sanger_data}
\title{Simulate sanger chromatogram density data}
\usage{
simulate_sanger_data(sequence, sd = 5, noise = 0.2, degrade = 0.8)
}
\arguments{
\item{sequence}{A string of nucleotides to use (GATC)}

\item{sd}{The standard deviation of the peak width}

\item{noise}{What proportion of the signal to make from random noise}

\item{degrade}{What proportion of the signal should be lost by the end}
}
\value{
A tibble with positions (20 per base) and G/A/T/C signal
}
\description{
Simulate sanger chromatogram density data
}
\examples{
simulate_sanger_data("GAATTC")
}
